Windows compatibility fixes
Windows paths can have an optional drive component, which complicates
things. We now use a custom `parse_path` function instead of directly
using `urlparse`, which handles this logic for us. This led to a new
utility module `prefect.utilities.filesystems`. I also moved the
`read_bytes_from_path` method there, and relocated the tests.
